Move .cmp reading into its own module
